# Improve the runtime performance of ViewFactor