@jmarrec 2019-10-10
Nice job @kbenne, but this lacks comments. We're doing unusual stuff here, so comments would be nice, added some in #3720